+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # 更新日志
 
+25w14t-2
+---
+- PRE-RELEASE: 此版本是测试验证版本,请勿在生产环境中使用;
+- CHANGE: 使用`touka-httpc`封装`HTTP Client`，更新到`v0.1.0`版本, 参看`touka-httpc`
+- CHANGE: 重构`whitelist`实现
+- CHANGE: 对`proxy`进行结构性调整
+- CHANGE: `chunckedreq`与`gitreq`共用`BufferPool`和`HTTP Client`
+- CHANGE: 新增`HTTP Client`配置块
+
 25w14t-1
 ---
 - PRE-RELEASE: 此版本是测试验证版本,请勿在生产环境中使用;